refactor(authentication): remove repository subdirectory
