+++ authentication/_mocks/gen.go
@@ -1,4 +1,4 @@
 package _mocks
 
-//go:generate mockgen -source ../accounts/repository/repository.go -destination=./repository_mocks/repository_mocks.go
+//go:generate mockgen -source ../accounts/repository.go -destination=./repository_mocks/repository_mocks.go
 //go:generate mockgen -source ../pwhash/pwhasher.go -destination=./pwhash_mocks/pwhash_mocks.go
